# internal/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the Nexodus API Server.",
        "title": "Nexodus API",
        "contact": {
            "name": "The Nexodus Authors",
            "url": "https://github.com/nexodus-io/nexodus/issues"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/vpc/{id}/events": {
            "post": {
                "description": "Watches events occurring in the vpc",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "VPC"
                ],
                "summary": "Watch events occurring in the vpc",
                "operationId": "WatchEvents",
                "parameters": [
                    {
                        "type": "string",
                        "description": "connect as the device with the given public key, device will be considered to be online for the duration of this request",
                        "name": "public_key",
                        "in": "query"
                    },
                    {
                        "description": "List of events to watch",
                        "name": "Watches",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Watch"
                            }
                        }
                    },
                    {
                        "type": "string",
                        "description": "VPC ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.WatchEvent"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.BaseError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.BaseError"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/models.BaseError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.InternalServerError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.BaseError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "something bad"
                }
            }
        },
...
        "models.InternalServerError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "something bad"
                },
                "trace_id": {
                    "type": "string"
                }
            }
        },
...
        "models.Watch": {
            "type": "object",
            "properties": {
                "at_tail": {
                    "type": "boolean"
                },
                "gt_revision": {
                    "type": "integer"
                },
                "kind": {
                    "type": "string"
                },
                "options": {
                    "type": "object",
                    "additionalProperties": true
                }
            }
        },
        "models.WatchEvent": {
            "type": "object",
            "properties": {
                "kind": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "value": {}
            }
        }
    },
    "securityDefinitions": {
        "OAuth2Implicit": {
            "type": "oauth2",
            "flow": "implicit",
            "authorizationUrl": "https://auth.try.nexodus.127.0.0.1.nip.io/",
            "scopes": {
                "admin": " Grants read and write access to administrative information",
                "user": " Grants read and write access to resources owned by this user"
            }
        }
    }
}